debug

+++ peer_pool.go
@@ -7,7 +7,6 @@ import (
 	"time"
 
 	"github.com/anacrolix/multiless"
-	"github.com/davecgh/go-spew/spew"
 	"github.com/google/btree"
 	"github.com/james-lawrence/torrent/internal/fnvx"
 )
@@ -20,7 +19,11 @@ type prioritizedPeer struct {
 }
 
 func priorityPeerCmp(a, b prioritizedPeer) bool {
-	return multiless.New().Bool(
+	return multiless.New().Uint64(
+		a.p.Attempts, b.p.Attempts, // prioritize peers we havent attempted before.
+	// ).Uint64(
+	// 	uint64(a.p.LastAttempt.UnixNano()), uint64(b.p.LastAttempt.UnixNano()), // prioritize peers we havent contacted recently
+	).Bool(
 		a.p.Trusted, b.p.Trusted,
 	).Uint32(
 		a.prio, b.prio,
@@ -34,8 +37,9 @@ func priorityPeerCmp(a, b prioritizedPeer) bool {
 func newPeerPool(n int, prio func(Peer) peerPriority) peerPool {
 	return peerPool{
 		m:         &sync.RWMutex{},
-		available: btree.NewG(n, priorityPeerCmp),
+		untried:   btree.NewG(n, priorityPeerCmp),
 		attempted: btree.NewG(n, priorityPeerCmp),
+		available: btree.NewG(n, priorityPeerCmp),
 		loaned:    make(map[netip.AddrPort]Peer, 32),
 		getPrio:   prio,
 		nextswap:  time.Now().Add(time.Minute),
@@ -44,8 +48,9 @@ func newPeerPool(n int, prio func(Peer) peerPriority) peerPool {
 
 type peerPool struct {
 	m         *sync.RWMutex
-	available *btree.BTreeG[prioritizedPeer]
+	untried   *btree.BTreeG[prioritizedPeer]
 	attempted *btree.BTreeG[prioritizedPeer]
+	available *btree.BTreeG[prioritizedPeer]
 	loaned    map[netip.AddrPort]Peer
 	getPrio   func(Peer) peerPriority
 	nextswap  time.Time
@@ -59,12 +64,17 @@ func (t *peerPool) Each(f func(Peer)) {
 	t.m.RLock()
 	defer t.m.RUnlock()
 
-	t.available.Ascend(func(item prioritizedPeer) bool {
+	t.untried.Ascend(func(item prioritizedPeer) bool {
 		f(item.p)
 		return true
 	})
 
 	t.attempted.Ascend(func(item prioritizedPeer) bool {
+		f(item.p)
+		return true
+	})
+
+	t.available.Ascend(func(item prioritizedPeer) bool {
 		f(item.p)
 		return true
 	})
@@ -77,13 +87,13 @@ func (t *peerPool) Each(f func(Peer)) {
 func (t *peerPool) Stats() (pending, connecting int) {
 	t.m.RLock()
 	defer t.m.RUnlock()
-	return t.available.Len(), len(t.loaned)
+	return t.untried.Len(), len(t.loaned)
 }
 
 func (t *peerPool) Len() int {
 	t.m.RLock()
 	defer t.m.RUnlock()
-	return t.available.Len()
+	return t.untried.Len()
 }
 
 func (t *peerPool) Connecting(p Peer) bool {
@@ -94,12 +104,14 @@ func (t *peerPool) Connecting(p Peer) bool {
 }
 
 // Peer is returned to the pool
-func (t *peerPool) Attempted(p Peer, c ConnStats) {
+func (t *peerPool) Attempted(p Peer, attempts uint64) {
 	t.m.Lock()
 	defer t.m.Unlock()
-
-	p.Attempts += 1
-	log.Println("returning", p.ID, p.AddrPort, p.Attempts, spew.Sdump(c))
+	ts := time.Now()
+	elapsed := time.Since(p.LastAttempt)
+	p.LastAttempt = ts
+	p.Attempts = attempts
+	log.Println("returning", elapsed, "untried", t.untried.Len(), "half", len(t.loaned), "available", t.available.Len(), "attempted", t.attempted.Len(), "attempts", p.Attempts, p.AddrPort)
 	delete(t.loaned, p.AddrPort)
 
 	t.attempted.ReplaceOrInsert(t.prioritized(p))
@@ -119,11 +131,19 @@ func (t *peerPool) Add(p Peer) bool {
 
 	prio := t.prioritized(p)
 
-	// if _, replaced := t.attempted.ReplaceOrInsert(prio); replaced {
-	// 	return replaced
-	// }
+	if exists := t.attempted.Has(prio); exists {
+		return exists
+	}
 
-	if _, replaced := t.available.ReplaceOrInsert(prio); replaced {
+	if exists := t.available.Has(prio); exists {
+		return exists
+	}
+
+	if _, exists := t.loaned[p.AddrPort]; exists {
+		return exists
+	}
+
+	if _, replaced := t.untried.ReplaceOrInsert(prio); replaced {
 		return replaced
 	}
 
@@ -134,12 +154,16 @@ func (t *peerPool) DeleteMin() (ret prioritizedPeer, ok bool) {
 	t.m.Lock()
 	defer t.m.Unlock()
 
-	return t.available.DeleteMin()
+	return t.untried.DeleteMin()
 }
 
 func (t *peerPool) PopMax() (p prioritizedPeer, ok bool) {
 	t.m.Lock()
 	defer t.m.Unlock()
+
+	if max, present := t.untried.DeleteMax(); present {
+		return max, present
+	}
 
 	if max, present := t.available.DeleteMax(); present {
 		return max, present
@@ -151,8 +175,15 @@ func (t *peerPool) PopMax() (p prioritizedPeer, ok bool) {
 		t.nextswap = ts.Add(time.Minute)
 	}
 
-	t.available = t.attempted.Clone()
-	t.attempted.Clear(true)
+	log.Println("returning moving failed to available")
+	for range 10 {
+		if item, ok := t.attempted.DeleteMax(); ok {
+			t.available.ReplaceOrInsert(item)
+			continue
+		}
+
+		break
+	}
 
 	return t.available.DeleteMax()
 }
